% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{new_family_tree}
\alias{new_family_tree}
\title{Core Functions}
\usage{
new_family_tree(
  fn = NULL,
  ln = NULL,
  sex = NULL,
  nn = NULL,
  yob = NULL,
  note = NULL
)
}
\arguments{
\item{fn}{first name}

\item{ln}{last name}

\item{sex}{sex}

\item{nn}{nick name or alternate name}

\item{yob}{year of birth}

\item{note}{other notes about a person}
}
\value{
a new family tree with one person
}
\description{
Core Functions
}
\examples{
new_family_tree(fn = 'Abraham', ln = 'Lincoln')
new_family_tree(fn = 'Abraham', ln = 'Lincoln', nn = 'Honest Abe', yob = 1809)

}
