% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{new_family_tree}
\alias{new_family_tree}
\title{Core Functions}
\usage{
new_family_tree(tree_name = "my new tree")
}
\arguments{
\item{tree_name}{a name for the tree to be created,
e.g. 'The Wallace Family'}
}
\value{
a new tree
}
\description{
Core Functions
}
\examples{
new_family_tree(tree_name = 'The Wallace Family')
}
