% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{new_family_tree}
\alias{new_family_tree}
\title{Core Functions}
\usage{
new_family_tree(
  first_name = NULL,
  last_name = NULL,
  sex = NULL,
  nick_name = NULL,
  year_of_birth = NULL,
  note = NULL
)
}
\arguments{
\item{first_name}{first name}

\item{last_name}{last name}

\item{sex}{sex}

\item{nick_name}{nick name or alternate name}

\item{year_of_birth}{year of birth}

\item{note}{other notes about a person}
}
\value{
a new family tree with one person
}
\description{
Core Functions
}
\examples{
new_family_tree(first_name = 'Abraham', last_name = 'Lincoln')
new_family_tree(first_name = 'Abraham', last_name = 'Lincoln',
                nick_name = 'Honest Abe', year_of_birth = 1809)

}
